+++ site-cookbooks/hhbd-admin/recipes/default.rb
@@ -1,8 +1,10 @@
 #
-# Cookbook Name:: hhbd-backend
+# Cookbook Name:: hhbd-admin
 # Recipe:: default
 #
 # Copyright 2015, YOUR_COMPANY_NAME
 #
 # All rights reserved - Do Not Redistribute
 #
+
+include_recipe "hhbd-admin::install"