+++ site-cookbooks/hhbd-content/attributes/default.rb
@@ -1,0 +1,1 @@
+node.default["hhbd-content"]["url"] = "s.hhbd.pl.vmx"