+++ site-cookbooks/hhbd-xadmin/attributes/default.rb
@@ -1,0 +1,4 @@
+node.default["app-xadmin"]["url"] = "xadmin.hhbd.pl.vmx"
+
+node.default['app-xadmin']['config'] = {}
+